+++ lib/easypost/exceptions/api/api_error.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require_relative 'easy_post_error'
+require_relative '../easy_post_error'
 require 'easypost/constants'
 
 class EasyPost::Exceptions::ApiError < EasyPost::Exceptions::EasyPostError
@@ -21,7 +21,7 @@ class EasyPost::Exceptions::ApiError < EasyPost::Exceptions::EasyPostError
     error_string
   end
 
-  def self.get_potential_error(response)
+  def self.handle_api_error(response)
     status_code = response.code
     status_code = status_code.to_i if status_code.is_a?(String)
 
@@ -44,7 +44,7 @@ class EasyPost::Exceptions::ApiError < EasyPost::Exceptions::EasyPostError
       errors = nil
     end
 
-    cls = EasyPost::Constants.exception_cls_from_status_code(status_code)
+    cls = exception_cls_from_status_code(status_code)
 
     if cls == EasyPost::Exceptions::UnknownHttpError
       error_message = EasyPost::Constants::ErrorMessages::UNEXPECTED_HTTP_STATUS_CODE % status_code
@@ -53,5 +53,38 @@ class EasyPost::Exceptions::ApiError < EasyPost::Exceptions::EasyPostError
 
     # Return (don't throw here) an instance of the appropriate error class
     cls.new(error_message, status_code, error_type, errors)
+  end
+
+  def self.exception_cls_from_status_code(status_code)
+    case status_code
+    when 0
+      EasyPost::Exceptions::ConnectionError
+    when 100, 101, 102, 103
+      EasyPost::Exceptions::UnknownHttpError
+    when 300, 301, 302, 303, 304, 305, 306, 307, 308
+      EasyPost::Exceptions::RedirectError
+    when 401, 403
+      EasyPost::Exceptions::UnauthorizedError
+    when 402
+      EasyPost::Exceptions::PaymentError
+    when 404
+      EasyPost::Exceptions::NotFoundError
+    when 405
+      EasyPost::Exceptions::MethodNotAllowedError
+    when 408
+      EasyPost::Exceptions::TimeoutError
+    when 422
+      EasyPost::Exceptions::InvalidRequestError
+    when 429
+      EasyPost::Exceptions::RateLimitError
+    when 500
+      EasyPost::Exceptions::InternalServerError
+    when 502, 504
+      EasyPost::Exceptions::GatewayTimeoutError
+    when 503
+      EasyPost::Exceptions::ServiceUnavailableError
+    else
+      EasyPost::Exceptions::UnknownHttpError
+    end
   end
 end
