+++ lib/easypost/exceptions/api/internal_server_error.rb
@@ -3,7 +3,4 @@
 require_relative 'api_error'
 
 class EasyPost::Exceptions::InternalServerError < EasyPost::Exceptions::ApiError
-  def initialize(message, status_code = nil, error_type = nil, errors = nil)
-    super message, status_code, error_type, errors
-  end
 end
